Split nzbstuff into separate files for Article, NzbFile and NzbObject
