+++ sabnzbd/nzb/object.py
@@ -16,7 +16,7 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 """
-sabnzbd.nzbstuff - misc
+sabnzbd.nzb.object - NzbObject class for representing NZB download jobs
 """
 import os
 import time
@@ -30,7 +30,8 @@ from typing import Any, Optional, Union, BinaryIO, Deque
 
 # SABnzbd modules
 import sabnzbd
-import sabctools
+from sabnzbd.nzb.article import TryList, Article, TRYLIST_LOCK
+from sabnzbd.nzb.file import NzbFile
 from sabnzbd.constants import (
     GIGI,
     ATTRIB_FILE,
@@ -60,6 +61,8 @@ from sabnzbd.misc import (
     opts_to_pp,
     pp_to_opts,
     duplicate_warning,
+    scan_password,
+    subject_name_extractor,
 )
 from sabnzbd.filesystem import (
     sanitize_foldername,
@@ -89,463 +92,19 @@ from sabnzbd.filesystem import (
     remove_data,
     strip_extensions,
     get_ext,
+    create_work_name,
+    nzf_cmp_name,
+    RAR_RE,
 )
 from sabnzbd.par2file import FilePar2Info, has_par2_in_filename, analyse_par2, parse_par2_file, is_par2_file
 from sabnzbd.decorators import synchronized
 import sabnzbd.config as config
 import sabnzbd.cfg as cfg
-import sabnzbd.nzbparser
 from sabnzbd.downloader import Server
 from sabnzbd.database import HistoryDB
 from sabnzbd.deobfuscate_filenames import is_probably_obfuscated
 
-# Name patterns
-# In the subject, we expect the filename within double quotes
-RE_SUBJECT_FILENAME_QUOTES = re.compile(r'"([^"]*)"')
-# Otherwise something that looks like a filename
-RE_SUBJECT_BASIC_FILENAME = re.compile(r"\b([\w\-+()' .,]+(?:\[[\w\-/+()' .,]*][\w\-+()' .,]*)*\.[A-Za-z0-9]{2,4})\b")
-RE_RAR = re.compile(r"(\.rar|\.r\d\d|\.s\d\d|\.t\d\d|\.u\d\d|\.v\d\d)$", re.I)
 
-
-##############################################################################
-# Trylist
-##############################################################################
-
-TRYLIST_LOCK = threading.RLock()
-
-
-class TryList:
-    """TryList keeps track of which servers have been tried for a specific article"""
-
-    # Pre-define attributes to save memory
-    __slots__ = ("try_list",)
-
-    def __init__(self):
-        # Sets are faster than lists
-        self.try_list: set[Server] = set()
-
-    def server_in_try_list(self, server: Server) -> bool:
-        """Return whether specified server has been tried"""
-        with TRYLIST_LOCK:
-            return server in self.try_list
-
-    def all_servers_in_try_list(self, all_servers: set[Server]) -> bool:
-        """Check if all servers have been tried"""
-        with TRYLIST_LOCK:
-            return all_servers.issubset(self.try_list)
-
-    def add_to_try_list(self, server: Server):
-        """Register server as having been tried already"""
-        with TRYLIST_LOCK:
-            # Sets cannot contain duplicate items
-            self.try_list.add(server)
-
-    def remove_from_try_list(self, server: Server):
-        """Remove server from list of tried servers"""
-        with TRYLIST_LOCK:
-            # Discard does not require the item to be present
-            self.try_list.discard(server)
-
-    def reset_try_list(self):
-        """Clean the list"""
-        with TRYLIST_LOCK:
-            self.try_list = set()
-
-    def __getstate__(self):
-        """Save the servers"""
-        return set(server.id for server in self.try_list)
-
-    def __setstate__(self, servers_ids: list[str]):
-        self.try_list = set()
-        for server in sabnzbd.Downloader.servers:
-            if server.id in servers_ids:
-                self.add_to_try_list(server)
-
-
-##############################################################################
-# Article
-##############################################################################
-ArticleSaver = (
-    "article",
-    "art_id",
-    "bytes",
-    "lowest_partnum",
-    "decoded",
-    "file_size",
-    "data_begin",
-    "data_size",
-    "on_disk",
-    "nzf",
-    "crc32",
-)
-
-
-class Article(TryList):
-    """Representation of one article"""
-
-    # Pre-define attributes to save memory
-    __slots__ = ArticleSaver + ("fetcher", "fetcher_priority", "tries")
-
-    def __init__(self, article, article_bytes, nzf):
-        super().__init__()
-        self.article: str = article
-        self.art_id: Optional[str] = None
-        self.bytes: int = article_bytes
-        self.lowest_partnum: bool = False
-        self.fetcher: Optional[Server] = None
-        self.fetcher_priority: int = 0
-        self.tries: int = 0  # Try count
-        self.decoded: bool = False
-        self.file_size: Optional[int] = None
-        self.data_begin: Optional[int] = None
-        self.data_size: Optional[int] = None
-        self.on_disk: bool = False
-        self.crc32: Optional[int] = None
-        self.nzf: NzbFile = nzf
-
-    @synchronized(TRYLIST_LOCK)
-    def reset_try_list(self):
-        """In addition to resetting the try list, also reset fetcher so all servers
-        are tried again. Locked so fetcher setting changes are also protected."""
-        self.fetcher = None
-        self.fetcher_priority = 0
-        super().reset_try_list()
-
-    @synchronized(TRYLIST_LOCK)
-    def allow_new_fetcher(self, remove_fetcher_from_try_list: bool = True):
-        """Let article get new fetcher and reset try lists of file and job.
-        Locked so all resets are performed at once"""
-        if remove_fetcher_from_try_list:
-            self.remove_from_try_list(self.fetcher)
-        self.fetcher = None
-        self.tries = 0
-        self.nzf.reset_try_list()
-        self.nzf.nzo.reset_try_list()
-
-    def get_article(self, server: Server, servers: list[Server]):
-        """Return article when appropriate for specified server"""
-        if self.fetcher or self.server_in_try_list(server):
-            return None
-
-        if server.priority > self.fetcher_priority:
-            # Check for higher priority server, taking advantage of servers list being sorted by priority
-            for server_check in servers:
-                if server_check.priority < server.priority:
-                    if server_check.active and not self.server_in_try_list(server_check):
-                        # There is a higher priority server, so set article priority and return
-                        self.fetcher_priority = server_check.priority
-                        return None
-                else:
-                    # All servers with a higher priority have been checked
-                    break
-
-        # If no higher priority servers, use this server
-        self.fetcher_priority = server.priority
-        self.fetcher = server
-        self.tries += 1
-        return self
-
-    def get_art_id(self):
-        """Return unique article storage name, create if needed"""
-        if not self.art_id:
-            self.art_id = get_new_id("article", self.nzf.nzo.admin_path)
-        return self.art_id
-
-    def search_new_server(self):
-        """Search for a new server for this article"""
-        # Since we need a new server, this one can be listed as failed
-        sabnzbd.BPSMeter.register_server_article_failed(self.fetcher.id)
-        self.add_to_try_list(self.fetcher)
-        # Servers-list could be modified during iteration, so we need a copy
-        for server in sabnzbd.Downloader.servers[:]:
-            if server.active and not self.server_in_try_list(server):
-                if server.priority >= self.fetcher.priority:
-                    self.tries = 0
-                    # Allow all servers for this nzo and nzf again (but not this fetcher for this article)
-                    self.allow_new_fetcher(remove_fetcher_from_try_list=False)
-                    return True
-
-        logging.info("Article %s unavailable on all servers, discarding", self.article)
-        return False
-
-    def __getstate__(self):
-        """Save to pickle file, selecting attributes"""
-        dict_ = {}
-        for item in ArticleSaver:
-            dict_[item] = getattr(self, item)
-        dict_["try_list"] = super().__getstate__()
-        return dict_
-
-    def __setstate__(self, dict_):
-        """Load from pickle file, selecting attributes"""
-        for item in ArticleSaver:
-            try:
-                setattr(self, item, dict_[item])
-            except KeyError:
-                # Handle new attributes
-                setattr(self, item, None)
-        super().__setstate__(dict_.get("try_list", []))
-        self.fetcher = None
-        self.fetcher_priority = 0
-        self.tries = 0
-
-    def __repr__(self):
-        return "<Article: article=%s, bytes=%s, art_id=%s>" % (self.article, self.bytes, self.art_id)
-
-
-##############################################################################
-# NzbFile
-##############################################################################
-class SkippedNzbFile(Exception):
-    pass
-
-
-NzbFileSaver = (
-    "date",
-    "filename",
-    "filename_checked",
-    "filepath",
-    "type",
-    "is_par2",
-    "vol",
-    "blocks",
-    "setname",
-    "articles",
-    "decodetable",
-    "bytes",
-    "bytes_left",
-    "nzo",
-    "nzf_id",
-    "deleted",
-    "import_finished",
-    "crc32",
-    "assembled",
-    "md5of16k",
-)
-
-
-class NzbFile(TryList):
-    """Representation of one file consisting of multiple articles"""
-
-    # Pre-define attributes to save memory
-    __slots__ = NzbFileSaver + ("lock",)
-
-    def __init__(self, date, subject, raw_article_db, file_bytes, nzo):
-        """Setup object"""
-        super().__init__()
-        self.lock = threading.RLock()
-
-        self.date: datetime.datetime = date
-        self.type: Optional[str] = None
-        self.filename: str = sanitize_filename(name_extractor(subject))
-        self.filename_checked = False
-        self.filepath: Optional[str] = None
-
-        # Identifiers for par2 files
-        self.is_par2: bool = False
-        self.vol: Optional[int] = None
-        self.blocks: Optional[int] = None
-        self.setname: Optional[str] = None
-
-        # Articles are removed from "articles" after being fetched
-        self.articles: dict[Article, Article] = {}
-        self.decodetable: list[Article] = []
-
-        self.bytes: int = file_bytes
-        self.bytes_left: int = file_bytes
-
-        self.nzo: NzbObject = nzo
-        self.deleted = False
-        self.import_finished = False
-
-        self.crc32: Optional[int] = 0
-        self.assembled: bool = False
-        self.md5of16k: Optional[bytes] = None
-
-        # Add first article to decodetable, this way we can check
-        # if this is maybe a duplicate nzf
-        if raw_article_db:
-            first_article = self.add_article(raw_article_db.pop(0))
-            first_article.lowest_partnum = True
-
-        if self in nzo.files:
-            logging.info("File %s occurred twice in NZB, skipping", self.filename)
-            raise SkippedNzbFile
-
-        # Create file on disk, which can fail in case of disk errors
-        self.nzf_id: str = get_new_id("nzf", nzo.admin_path)
-        if not self.nzf_id:
-            # Error already shown to user from get_new_id
-            raise SkippedNzbFile
-
-        # Any articles left?
-        if raw_article_db:
-            # Save the rest
-            save_data(raw_article_db, self.nzf_id, nzo.admin_path)
-        else:
-            # All imported
-            self.import_finished = True
-
-    def finish_import(self):
-        """Load the article objects from disk"""
-        logging.debug("Finishing import on %s", self.filename)
-        if raw_article_db := load_data(self.nzf_id, self.nzo.admin_path, remove=False):
-            for raw_article in raw_article_db:
-                self.add_article(raw_article)
-
-            # Make sure we have labeled the lowest part number
-            # Also when DirectUnpack is disabled we need to know
-            self.decodetable[0].lowest_partnum = True
-
-            # Mark safe to continue
-            self.import_finished = True
-
-    def add_article(self, article_info):
-        """Add article to object database and return article object"""
-        article = Article(article_info[0], article_info[1], self)
-        with self.lock:
-            self.articles[article] = article
-            self.decodetable.append(article)
-        return article
-
-    def remove_article(self, article: Article, success: bool) -> int:
-        """Handle completed article, possibly end of file"""
-        with self.lock:
-            if self.articles.pop(article, None) is not None:
-                if success:
-                    self.bytes_left -= article.bytes
-            return len(self.articles)
-
-    def set_par2(self, setname, vol, blocks):
-        """Designate this file as a par2 file"""
-        self.is_par2 = True
-        self.setname = setname
-        self.vol = vol
-        self.blocks = int_conv(blocks)
-
-    def update_crc32(self, crc32: Optional[int], length: int) -> None:
-        if self.crc32 is None or crc32 is None:
-            self.crc32 = None
-        else:
-            self.crc32 = sabctools.crc32_combine(self.crc32, crc32, length)
-
-    def get_articles(self, server: Server, servers: list[Server], fetch_limit: int):
-        """Get next articles to be downloaded"""
-        articles = server.article_queue
-        with self.lock:
-            for article in self.articles:
-                if article := article.get_article(server, servers):
-                    articles.append(article)
-                    if len(articles) >= fetch_limit:
-                        return
-        self.add_to_try_list(server)
-
-    @synchronized(TRYLIST_LOCK)
-    def reset_all_try_lists(self):
-        """Reset all try lists. Locked so reset is performed
-        for all items at the same time without chance of another
-        thread changing any of the items while we are resetting"""
-        with self.lock:
-            for art in self.articles:
-                art.reset_try_list()
-        self.reset_try_list()
-
-    def first_article_processed(self) -> bool:
-        """Check if the first article has been processed.
-        This ensures we have attempted to extract md5of16k and filename information
-        before creating the filepath.
-        """
-        # The first article of decodetable is always the lowest
-        first_article = self.decodetable[0]
-        # If it's still in nzo.first_articles, it hasn't been processed yet
-        return first_article not in self.nzo.first_articles
-
-    def prepare_filepath(self):
-        """Do all checks before making the final path"""
-        if not self.filepath:
-            # Wait for the first article to be processed so we can get md5of16k
-            # and proper filename before creating the filepath
-            if not self.first_article_processed():
-                return None
-
-            self.nzo.verify_nzf_filename(self)
-            filename = sanitize_filename(self.filename)
-            self.filepath = get_unique_filename(os.path.join(self.nzo.download_path, filename))
-            self.filename = get_filename(self.filepath)
-        return self.filepath
-
-    @property
-    def completed(self):
-        """Is this file completed?"""
-        if not self.import_finished:
-            return False
-        with self.lock:
-            return not self.articles
-
-    def remove_admin(self):
-        """Remove article database from disk (sabnzbd_nzf_<id>)"""
-        try:
-            logging.debug("Removing article database for %s", self.nzf_id)
-            remove_file(os.path.join(self.nzo.admin_path, self.nzf_id))
-        except Exception:
-            pass
-
-    def __enter__(self):
-        self.lock.acquire()
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        self.lock.release()
-
-    def __getstate__(self):
-        """Save to pickle file, selecting attributes"""
-        dict_ = {}
-        for item in NzbFileSaver:
-            dict_[item] = getattr(self, item)
-        dict_["try_list"] = super().__getstate__()
-        return dict_
-
-    def __setstate__(self, dict_):
-        """Load from pickle file, selecting attributes"""
-        for item in NzbFileSaver:
-            try:
-                setattr(self, item, dict_[item])
-            except KeyError:
-                # Handle new attributes
-                setattr(self, item, None)
-        super().__setstate__(dict_.get("try_list", []))
-        self.lock = threading.RLock()
-        if isinstance(self.articles, list):
-            # Converted from list to dict
-            self.articles = {x: x for x in self.articles}
-
-    def __eq__(self, other: "NzbFile"):
-        """Assume it's the same file if the number bytes and first article
-        are the same or if there are no articles left, use the filenames.
-        Some NZB's are just a mess and report different sizes for the same article.
-        We used to compare (__eq__) articles based on article-ID, however, this failed
-        because some NZB's had the same article-ID twice within one NZF.
-        """
-        if other and (self.bytes == other.bytes or len(self.decodetable) == len(other.decodetable)):
-            if self.decodetable and other.decodetable:
-                return self.decodetable[0].article == other.decodetable[0].article
-            # Fallback to filename comparison
-            return self.filename == other.filename
-        return False
-
-    def __hash__(self):
-        """Required because we implement eq. The same file can be spread
-        over multiple NZO's so we make every NZF unique. Even though
-        it's considered bad practice.
-        """
-        return id(self)
-
-    def __repr__(self):
-        return "<NzbFile: filename=%s, bytes=%s, nzf_id=%s>" % (self.filename, self.bytes, self.nzf_id)
-
-
-##############################################################################
-# NzbObject
-##############################################################################
 class NzbEmpty(Exception):
     pass
 
@@ -1032,7 +591,7 @@ class NzbObject(TryList):
             logging.debug("Unwanted Extension: putting last rar after first rar")
             firstrarpos = lastrarpos = 0
             for nzfposcounter, nzf in enumerate(self.files):
-                if RE_RAR.search(nzf.filename.lower()):
+                if RAR_RE.search(nzf.filename.lower()):
                     # a NZF found with '.rar' in the name
                     if firstrarpos == 0:
                         # this is the first .rar found, so remember this position
@@ -2134,109 +1693,3 @@ class NzbObject(TryList):
 
     def __repr__(self):
         return "<NzbObject: filename=%s, bytes=%s, nzo_id=%s>" % (self.filename, self.bytes, self.nzo_id)
-
-
-def nzf_cmp_name(nzf1: NzbFile, nzf2: NzbFile):
-    # The comparison will sort .par2 files to the top of the queue followed by .rar files,
-    # they will then be sorted by name.
-    nzf1_name = nzf1.filename.lower()
-    nzf2_name = nzf2.filename.lower()
-
-    # Determine vol-pars
-    is_par1 = ".vol" in nzf1_name and ".par2" in nzf1_name
-    is_par2 = ".vol" in nzf2_name and ".par2" in nzf2_name
-
-    # mini-par2 in front
-    if not is_par1 and nzf1_name.endswith(".par2"):
-        return -1
-    if not is_par2 and nzf2_name.endswith(".par2"):
-        return 1
-
-    # vol-pars go to the back
-    if is_par1 and not is_par2:
-        return 1
-    if is_par2 and not is_par1:
-        return -1
-
-    # Prioritize .rar files above any other type of file (other than vol-par)
-    m1 = RE_RAR.search(nzf1_name)
-    m2 = RE_RAR.search(nzf2_name)
-    if m1 and not (is_par2 or m2):
-        return -1
-    elif m2 and not (is_par1 or m1):
-        return 1
-    # Force .rar to come before 'r00'
-    if m1 and m1.group(1) == ".rar":
-        nzf1_name = nzf1_name.replace(".rar", ".r//")
-    if m2 and m2.group(1) == ".rar":
-        nzf2_name = nzf2_name.replace(".rar", ".r//")
-    return cmp(nzf1_name, nzf2_name)
-
-
-def create_work_name(name: str) -> str:
-    """Remove ".nzb" and ".par(2)" and sanitize, skip URL's"""
-    if name.find("://") < 0:
-        # Invalid charters need to be removed before and after (see unit-tests)
-        return sanitize_foldername(strip_extensions(sanitize_foldername(name)))
-    else:
-        return name.strip()
-
-
-def scan_password(name: str) -> tuple[str, Optional[str]]:
-    """Get password (if any) from the title"""
-    if "http://" in name or "https://" in name:
-        return name, None
-
-    # Strip any unwanted usenet-related extensions
-    name = strip_extensions(name)
-
-    # Identify any braces
-    braces = name[1:].find("{{")
-    if braces < 0:
-        braces = len(name)
-    else:
-        braces += 1
-    slash = name.find("/")
-
-    # Look for name/password, but make sure that '/' comes before any {{
-    if 0 < slash < braces and "password=" not in name:
-        # Is it maybe in 'name / password' notation?
-        if slash == name.find(" / ") + 1 and name[: slash - 1].strip(". "):
-            # Remove the extra space after name and before password
-            return name[: slash - 1].strip(". "), name[slash + 2 :]
-        if name[:slash].strip(". "):
-            return name[:slash].strip(". "), name[slash + 1 :]
-
-    # Look for "name password=password"
-    pw = name.find("password=")
-    if pw > 0 and name[:pw].strip(". "):
-        return name[:pw].strip(". "), name[pw + 9 :]
-
-    # Look for name{{password}}
-    if braces < len(name):
-        closing_braces = name.rfind("}}")
-        if closing_braces > braces and name[:braces].strip(". "):
-            return name[:braces].strip(". "), name[braces + 2 : closing_braces]
-
-    # Look again for name/password
-    if slash > 0 and name[:slash].strip(". "):
-        return name[:slash].strip(". "), name[slash + 1 :]
-
-    # No password found
-    return name, None
-
-
-def name_extractor(subject: str) -> str:
-    """Try to extract a file name from a subject line, return `subject` if in doubt"""
-    # Filename nicely wrapped in quotes
-    for name in re.findall(RE_SUBJECT_FILENAME_QUOTES, subject):
-        if name := name.strip(' "'):
-            return name
-
-    # Found nothing? Try a basic filename-like search
-    for name in re.findall(RE_SUBJECT_BASIC_FILENAME, subject):
-        if name := name.strip():
-            return name
-
-    # Return the subject
-    return subject
